Rename User.py to user.py
