+++ app/rewards/api.py
@@ -8,7 +8,7 @@ from app.auth import with_auth_user
 from app.events.event import Event
 from app.user.service import find_user_by_referral_code, find_user, update_user
 from app.user.errors import UserNotFoundError
-from app.user.User import User
+from app.user.user import User
 from .errors import ReferralCodeNotFound, InvalidReferralOperation
 
 POINTS_FOR_REFERRAL = 200
